requirements.txt
python-telegram-bot
telepot == 6.7
gTTS
